Removed dead code and moved Scheduler internal functions to Traits
